+++ sumologic/resource_sumologic_cse_match_list_test.go
@@ -13,10 +13,15 @@ func TestAccSumologicSCEMatchList_create(t *testing.T) {
 
 	var matchList CSEMatchListGet
 	nActive := true
-	nDefaultTtl := 1
-	nDescription := "New Test Status Description"
+	nDefaultTtl := 10800
+	nDescription := "New Match List Description"
 	nName := "Match List Name"
 	nTargetColumn := "SrcIp"
+	liActive := true
+	liDescription := "Match List Item Description"
+	liValue := "value"
+	liExpiration := "2122-02-27T04:00:00"
+
 	resourceName := "sumologic_cse_match_list.match_list"
 	resource.Test(t, resource.TestCase{
 		PreCheck:     func() { testAccPreCheck(t) },
@@ -24,11 +29,53 @@ func TestAccSumologicSCEMatchList_create(t *testing.T) {
 		CheckDestroy: testAccCSEMatchListDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testCreateCSEMatchListConfig(nActive, nDefaultTtl, nDescription, nName, nTargetColumn),
+				Config: testCreateCSEMatchListConfig(nActive, nDefaultTtl, nDescription, nName, nTargetColumn, liActive, liDescription, liExpiration, liValue),
 				Check: resource.ComposeTestCheckFunc(
 					testCheckCSEMatchListExists(resourceName, &matchList),
 					testCheckMatchListValues(&matchList, nDefaultTtl, nDescription, nName, nTargetColumn),
 					resource.TestCheckResourceAttrSet(resourceName, "id"),
+				),
+			},
+		},
+	})
+}
+
+func TestAccSumologicSCEMatchList_update(t *testing.T) {
+	SkipCseTest(t)
+
+	var matchList CSEMatchListGet
+	nActive := true
+	nDefaultTtl := 10800
+	nDescription := "New Match List Description"
+	nName := "Match List Name"
+	nTargetColumn := "SrcIp"
+	liActive := true
+	liDescription := "Match List Item Description"
+	liValue := "value"
+	liExpiration := "2122-02-27T04:00:00"
+	uDefaultTtl := 3600
+	uDescription := "Updated Match List Description"
+	uliDescription := "Updated Match List item Description"
+	resourceName := "sumologic_cse_match_list.match_list"
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCSEMatchListDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testCreateCSEMatchListConfig(nActive, nDefaultTtl, nDescription, nName, nTargetColumn, liActive, liDescription, liExpiration, liValue),
+				Check: resource.ComposeTestCheckFunc(
+					testCheckCSEMatchListExists(resourceName, &matchList),
+					testCheckMatchListValues(&matchList, nDefaultTtl, nDescription, nName, nTargetColumn),
+					resource.TestCheckResourceAttrSet(resourceName, "id"),
+				),
+			},
+			{
+				Config: testCreateCSEMatchListConfig(nActive, uDefaultTtl, uDescription, nName, nTargetColumn, liActive, uliDescription, liExpiration, liValue),
+				Check: resource.ComposeTestCheckFunc(
+					testCheckCSEMatchListExists(resourceName, &matchList),
+					testCheckMatchListValues(&matchList, uDefaultTtl, uDescription, nName, nTargetColumn),
 				),
 			},
 		},
@@ -58,7 +105,7 @@ func testAccCSEMatchListDestroy(s *terraform.State) error {
 	return nil
 }
 
-func testCreateCSEMatchListConfig(nActive bool, nDefaultTtl int, nDescription string, nName string, nTargetColumn string) string {
+func testCreateCSEMatchListConfig(nActive bool, nDefaultTtl int, nDescription string, nName string, nTargetColumn string, liActive bool, liDescription string, liExpiration string, liValue string) string {
 	return fmt.Sprintf(`
 resource "sumologic_cse_match_list" "match_list" {
 	active = "%t"
@@ -66,8 +113,14 @@ resource "sumologic_cse_match_list" "match_list" {
 	description = "%s"
 	name = "%s"
 	target_column = "%s"
+	items {
+		active = "%t"
+		description = "%s"
+		expiration = "%s"
+		value = "%s"
+	}
 }
-`, nActive, nDefaultTtl, nDescription, nName, nTargetColumn)
+`, nActive, nDefaultTtl, nDescription, nName, nTargetColumn, liActive, liDescription, liExpiration, liValue)
 }
 
 func testCheckCSEMatchListExists(n string, matchList *CSEMatchListGet) resource.TestCheckFunc {
@@ -80,7 +133,6 @@ func testCheckCSEMatchListExists(n string, matchList *CSEMatchListGet) resource.
 		if rs.Primary.ID == "" {
 			return fmt.Errorf("match List ID is not set")
 		}
-
 		c := testAccProvider.Meta().(*Client)
 		matchListResp, err := c.GetCSEMatchList(rs.Primary.ID)
 		if err != nil {
@@ -88,6 +140,29 @@ func testCheckCSEMatchListExists(n string, matchList *CSEMatchListGet) resource.
 		}
 
 		*matchList = *matchListResp
+
+		return nil
+	}
+}
+
+func testCheckCSEMatchListItemExists(n string, matchListItem *CSEMatchListItemGet) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		rs, ok := s.RootModule().Resources[n]
+		if !ok {
+			return fmt.Errorf("not found: %s", n)
+		}
+
+		if rs.Primary.ID == "" {
+			return fmt.Errorf("match List ID is not set")
+		}
+
+		c := testAccProvider.Meta().(*Client)
+		matchListResp, err := c.GetCSEMatchListItem(rs.Primary.ID)
+		if err != nil {
+			return err
+		}
+
+		*matchListItem = *matchListResp
 
 		return nil
 	}